Added frontend UI

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc Get goals
 // @route Get  /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     
     res.status(200).json(goals)
 })
@@ -20,7 +21,8 @@ const setGoal = asyncHandler(async (req, res) => {
       throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
         })
 
     res.status(200).json(goal)
@@ -36,6 +38,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+       res.status(401)
+       throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if(user.id !== goal.user.toString()) {
+        res.status(401)
+        throw new Error('User not authorized')
+        }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.
         body, {
@@ -55,6 +71,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+       res.status(401)
+       throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if(user.id !== goal.user.toString()) {
+        res.status(401)
+        throw new Error('User not authorized')
+        }
 
     await goal.deleteOne()
